Update default solution and record cry duration

+++ Assets/Scripts/Managers/ScoreboardManager.cs
@@ -78,7 +78,8 @@ public class ScoreboardManager : MonoBehaviour
 
     public void SetCryDuration()
     {
-        cryDurationText.text = gameStats.totalCryingDuration.ToString() + " seconds";
+
+        cryDurationText.text = ((int)gameStats.totalCryingDuration).ToString() + " seconds";
     }
 
     public void SetDiaperCount()
